refactor: 책 통계 조회 DB로 count 조회하도록 수정

+++ src/main/kotlin/com/group/libraryapp/dto/book/response/BookStatResponse.kt
@@ -4,5 +4,5 @@ import com.group.libraryapp.service.book.BookType
 
 class BookStatResponse(
     val type: BookType,
-    val count: Int
+    val count: Long
 )
